+++ tests/drmcp/unit/predictive_tools/test_data.py
@@ -13,9 +13,7 @@
 # limitations under the License.
 
 import base64
-from collections.abc import Iterator
 from unittest.mock import MagicMock
-from unittest.mock import Mock
 from unittest.mock import patch
 
 import datarobot as dr
@@ -24,19 +22,9 @@ import pytest
 from datarobot.errors import ClientError
 from datarobot.models.data_store import DataStoreParameters
 
-from datarobot_genai.drtools.core.clients.datarobot import ThreadSafeDataRobotClient
 from datarobot_genai.drtools.core.exceptions import ToolError
 from datarobot_genai.drtools.core.exceptions import ToolErrorKind
 from datarobot_genai.drtools.predictive import data
-
-
-@pytest.fixture
-def mock_get_client_context_with_token_from_request_header() -> Iterator[Mock]:
-    with patch.object(
-        ThreadSafeDataRobotClient,
-        "get_client_context_with_token_from_request_header",
-    ) as mock_func:
-        yield mock_func
 
 
 def test_merge_pagination_metadata_adds_offset_limit_next_total() -> None:
