+++ tests/drmcp/unit/predictive_tools/test_deployment.py
@@ -15,10 +15,8 @@
 import os
 import shutil
 import tempfile
-from collections.abc import Iterator
 from pathlib import Path
 from unittest.mock import MagicMock
-from unittest.mock import Mock
 from unittest.mock import patch
 
 import datarobot as dr
@@ -35,15 +33,6 @@ from datarobot_genai.drtools.predictive import deployment
 
 def test_load_dotenv() -> None:
     load_dotenv(verbose=True)
-
-
-@pytest.fixture
-def mock_get_client_context_with_token_from_request_header() -> Iterator[Mock]:
-    with patch.object(
-        ThreadSafeDataRobotClient,
-        "get_client_context_with_token_from_request_header",
-    ) as mock_func:
-        yield mock_func
 
 
 @pytest.mark.asyncio
