+++ tests/drmcp/unit/predictive_tools/test_deployment_info.py
@@ -13,11 +13,9 @@
 # limitations under the License.
 
 import json
-from collections.abc import Iterator
 from typing import Any
 from unittest.mock import AsyncMock
 from unittest.mock import MagicMock
-from unittest.mock import Mock
 from unittest.mock import patch
 
 import datarobot as dr
@@ -26,7 +24,6 @@ import pytest
 from fastmcp.tools.tool import ToolResult
 from mcp.types import TextContent
 
-from datarobot_genai.drtools.core.clients.datarobot import ThreadSafeDataRobotClient
 from datarobot_genai.drtools.core.exceptions import ToolError
 from datarobot_genai.drtools.predictive.deployment_info import generate_prediction_data_template
 from datarobot_genai.drtools.predictive.deployment_info import get_deployment_features
@@ -34,15 +31,6 @@ from datarobot_genai.drtools.predictive.deployment_info import get_deployment_in
 from datarobot_genai.drtools.predictive.deployment_info import validate_prediction_data
 
 IMPORTANCE_THRESHOLD_TEST = 0.8
-
-
-@pytest.fixture
-def mock_get_client_context_with_token_from_request_header() -> Iterator[Mock]:
-    with patch.object(
-        ThreadSafeDataRobotClient,
-        "get_client_context_with_token_from_request_header",
-    ) as mock_func:
-        yield mock_func
 
 
 def _extract_content(result_obj: dict | ToolResult | ToolError) -> str:
