+++ tests/drmcp/unit/predictive_tools/test_model.py
@@ -13,7 +13,6 @@
 # limitations under the License.
 
 import json
-from collections.abc import Iterator
 from unittest.mock import MagicMock
 from unittest.mock import Mock
 from unittest.mock import patch
@@ -29,15 +28,6 @@ from datarobot_genai.drtools.core.exceptions import ToolErrorKind
 from datarobot_genai.drtools.predictive import model
 from datarobot_genai.drtools.predictive.model import ModelEncoder
 from datarobot_genai.drtools.predictive.model import model_to_dict
-
-
-@pytest.fixture
-def mock_get_client_context_with_token_from_request_header() -> Iterator[Mock]:
-    with patch.object(
-        ThreadSafeDataRobotClient,
-        "get_client_context_with_token_from_request_header",
-    ) as mock_func:
-        yield mock_func
 
 
 @pytest.mark.asyncio
