+++ tests/drmcp/unit/predictive_tools/test_training.py
@@ -12,27 +12,15 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from collections.abc import Iterator
 from unittest.mock import MagicMock
-from unittest.mock import Mock
 from unittest.mock import patch
 
 import datarobot as dr
 import pandas as pd
 import pytest
 
-from datarobot_genai.drtools.core.clients.datarobot import ThreadSafeDataRobotClient
 from datarobot_genai.drtools.core.exceptions import ToolError
 from datarobot_genai.drtools.predictive import training
-
-
-@pytest.fixture
-def mock_get_client_context_with_token_from_request_header() -> Iterator[Mock]:
-    with patch.object(
-        ThreadSafeDataRobotClient,
-        "get_client_context_with_token_from_request_header",
-    ) as mock_func:
-        yield mock_func
 
 
 @pytest.mark.asyncio
